Add run subcommand
commit-id:c3eaac89
